requirements.txt
pymupdf>=1.24.0
numpy>=1.26.0
transformers>=4.40.0
accelerate>=0.30.0
torch>=2.3.0 --index-url https://download.pytorch.org/whl/cu121